% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorRampAlpha.R
\name{colorRampAlpha}
\alias{colorRampAlpha}
\title{color ramp palette with transparency}
\usage{
colorRampAlpha(..., n, alpha)
}
\arguments{
\item{...}{arguments to pass to \code{\link[grDevices]{colorRampPalette}}.}

\item{n}{number of colors in palette}

\item{alpha}{alpha channel (opacity) value [0;1]}
}
\value{
colorRampAlpha returns a vector of defined colors with transparency interpolating the given sequence.
}
\description{
color ramp palette with transparency
}
\details{
for details see \code{\link[grDevices]{colorRamp}} and \code{\link[grDevices]{colorRampPalette}}
}
\examples{
library(RColorBrewer)
colorRampAlpha(brewer.pal(9,"PuOr"), n = 10, alpha=.5)
}
