% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errormeas.R
\name{errormeas}
\alias{errormeas}
\title{bundle of error measures}
\usage{
errormeas(obs, sim)
}
\arguments{
\item{obs}{observation time series or vector}

\item{sim}{observation time series or vector}
}
\value{
named vector of error meassures
}
\description{
error measures comparing two datasets (simulation vs. observation).
}
\details{
r: Pearson correlation
rmse: Root Mean Square Error
rsd: ratio of standard deviation
nme: Normalized mean error
qu05.5%: 05th percentile statistical measure
qu95.95%: 95th percentile statistical measure
KGE: Kling-Gupta efficiency
}
\examples{
errormeas(1:10,1:10)
}
\seealso{
\code{\link[hydroGOF]{KGE}}
}
