% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_2EVE.R
\name{help_2EVE}
\alias{help_2EVE}
\title{Allocate path to root directories at EVE server.}
\usage{
help_2EVE(dir = 1, setwdir = TRUE)
}
\arguments{
\item{dir}{directory at EVE serever, 1: home, 2: work, 3: data}

\item{setwdir}{boolean, if TRUE set working directory to dir}
}
\value{
path to directory
}
\description{
Allocate path to root directories (/home, /work, /data) at EVE server.
}
\examples{

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
