% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_aggrLE.R
\name{help_aggrLE}
\alias{help_aggrLE}
\title{Aggregate latent heat data.}
\usage{
help_aggrLE(LE, perc_thresh = 0.1, ET = TRUE)
}
\arguments{
\item{LE}{zoo object, subdaily latent heat data in W/m2}

\item{perc_thresh}{threshold (percentage of NAs [0;1]) defining if specific day fulfills data quality}

\item{ET}{boolean, if TRUE latent heat (W/m2) will be converted to evapotranspiration (mm/d)}
}
\value{
zoo object, containing quality daily latent heat or evapotranspiration data
}
\description{
Daily aggregation of latent heat data from eddy covariance systems. Native data time resolution is sub-daily. Conversion to evapotranspiration flux (mm/d) is possible.
}
\examples{
library(zoo)
library(chron)
# load eddy covariance data
data(eses1)
# extract latent heat
LE <- eses1$ufz_lE
# spline interpolation
LE_spline <- zoo::na.spline(LE, maxgap=10)
# aggregate latent heat, convert to evapotranspiration
ET <- help_aggrLE(LE = LE_spline, perc_thresh = .1, ET = TRUE)
# plot time series
plot(ET)

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
