% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_aggrP.R
\name{help_aggrP}
\alias{help_aggrP}
\title{Aggregate precipitation data.}
\usage{
help_aggrP(P, perc_thresh = 0.1)
}
\arguments{
\item{P}{zoo object, subdaily precipitation data in mm}

\item{perc_thresh}{threshold (percentage of NAs [0;1]) defining if specific day fulfills data quality}
}
\value{
zoo object, containing quality daily precipitation data
}
\description{
Daily aggregation of precipitation data. Native data time resolution is sub-daily.
}
\examples{
library(zoo)
library(chron)
# load eddy covariance data
data(eses1)
# extract precipitation
P <- eses1$Precip
# aggregate precipitation, convert to daily sum
P <- help_aggrP(LE = P, perc_thresh = .1)
# plot time series
plot(P)

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
