% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_aggrTa.R
\name{help_aggrT}
\alias{help_aggrT}
\title{Aggregate air temperature data.}
\usage{
help_aggrT(Ta, perc_thresh = 0.1)
}
\arguments{
\item{perc_thresh}{threshold (percentage of NAs [0;1]) defining if specific day fulfills data quality}

\item{P}{zoo object, subdaily precipitation data in mm}

\item{fun}{aggregation function, e.g. min, max, mean, etc.}
}
\value{
zoo object, containing quality daily air temperature data
}
\description{
Daily aggregation of air temperature data to mean, minimum and maximum air temperature. 
Native data time resolution is sub-daily.
}
\examples{
library(zoo)
library(chron)
# load eddy covariance data
data(eses1)
# extract precipitation
Ta <- eses1$Ta
# aggregate air temperature, convert to min, max, mean air temperature
Ta <- help_aggrTa(Ta = Ta, perc_thresh = .1, fun=c(min, mean, max))
# plot time series
plot(Ta)

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
