% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_loadCarboEU.R
\name{help_loadCarboEU}
\alias{help_loadCarboEU}
\title{Load CarboEurope station data.}
\usage{

  help_loadCarboEU(datapath = "/Users/brennerj/tmp/eve_f2_data/data/processed/eddy_carbo_europe_201709/",
  station = "ESES1", stationlist = FALSE)
}
\arguments{
\item{datapath}{full path to pre-processed data files}

\item{station}{station descrption, e.g. "ESES1"}

\item{stationlist}{gather information on FLUXnet sites, if FALSE normal proceeding of function, if charater vector or NA site information output, e.g. c("PT","ES") for sites over Iberian peninsula.}
}
\value{
zoo object, containing all CarboEU station data, chron datetime representation
}
\description{
Load CarboEurope station data.
}
\examples{

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
