% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_ncaddproj.R
\name{help_ncaddproj}
\alias{help_ncaddproj}
\title{Add projected coordinates to netCDF.}
\usage{
help_ncaddproj(nc, to = "proj", coordsys = 23030, varnames = c(x = "lon",
  y = "lat"))
}
\arguments{
\item{nc}{netCDF files.}

\item{to}{chracter, if "degree": convert from pojected to lat/lon (degree); if "proj" convert from lat/lon to projected (meter).}

\item{coordsys}{integer, EPSG coordinate system definition, default 23030.}

\item{varnames}{named character vector containing variable names, either for lat/lon or projected coordinates, default: c(x="lon",y="lat")}
}
\value{
NULL, saves netCDF file containing projected and lat/lon coordinates
}
\description{
Add projected coordinates (in meter) to netCDF, by converting lat/lon (degree) representation.
}
\examples{

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
