% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_ncaddxy.R
\name{help_ncaddxy}
\alias{help_ncaddxy}
\title{Add projected coordinates to netCDF.}
\usage{
help_ncaddxy(nc, header = "header.txt", InvP = FALSE)
}
\arguments{
\item{nc}{netCDF files.}

\item{header}{header file name, default="header.txt".}

\item{InvP}{boolean, if TRUE reverse yy variable}
}
\value{
NULL, saves netCDF file containing x(x) and y(y) variables.
}
\description{
Add one-dimensional projected coordinates x(x) and y(y) (in meter) to netCDF using information from header.txt.
}
\examples{

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
