% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_nearestNeighbourInt.R
\name{help_nearestNeighbourInt}
\alias{help_nearestNeighbourInt}
\title{Nearest neighbour interpolation.}
\usage{
help_nearestNeighbourInt(dat)
}
\arguments{
\item{dat}{numeric vector}
}
\value{

}
\description{
Nearest neighbour interpolation
}
\examples{
library(zoo)
library(chron)
# load eddy covariance data
data(eses1)
# extract precipitation
P <- eses1$Precip
# nearest neighbor interpolation
P_interpol <- help_nearestNeighbourInt(P)

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
