% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_parseDateTime.R
\name{help_parseDateTime}
\alias{help_parseDateTime}
\title{parse to Date or Date.Time objects.}
\usage{
help_parseDateTime(x.date, x.time, format = list(date = "\%Y-\%m-\%d",
  time = "\%H:\%M:\%S"), parse.to = "chron", fasttime = FALSE,
  timezone = "GMT")
}
\arguments{
\item{x.date}{date character vector.}

\item{x.time}{time character vector.}

\item{format}{list containing datetime format, default: list(date=..., time=...), 
if format$time = NULL date representation (as.Date) will be used for parsing.}

\item{parse.to}{specifying datetime representation, that will be used for parsing, 
either chron (default) or as.POSIXct.}

\item{timezone}{timezone only necessary if POSIX representation used, default "GMT"}
}
\value{
date or date.time object
}
\description{
parse character vector to Date (base::as.Date) or Date.Time (base::as.POSIXct or chron::chron) objects.
}
\examples{
x.time <- c("00:20:00", "00:40:00", "01:00:00")
x.date <- rep("2012-01-01", 3)

# chron representation
help_parseDateTime(x.date, x.time, parse.to="chron")

# POSIX representation
help_parseDateTime(x.date, x.time, parse.to="POSIX",
                   timezone="GMT")

# Date representation
help_parseDateTime(x.date, x.time=NULL)

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
