% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_plotNClayer.R
\name{help_plotNClayer}
\alias{help_plotNClayer}
\title{plot single layer netCDF.}
\usage{
help_plotNClayer(infile, var = "pre", shp = NA, proj_utm = FALSE,
  crs_proj4 = "+proj=utm +zone=30 +ellps=intl +units=m +no_defs",
  plot_proj_utm = TRUE, outfolder,
  colorRampPal = colorRampPalette(brewer.pal(9, "Blues"))(100),
  zlims = NULL, xlims = NULL, ylims = NULL, flip = FALSE)
}
\arguments{
\item{infile}{netCDF file.}

\item{var}{netCDF variable.}

\item{shp}{shape file, optional.}

\item{proj_utm}{boolean, is ncfile projected on UTM}

\item{crs_proj4}{character, proj4 projection definition}

\item{plot_proj_utm}{boolean , should out plot be projected}

\item{outfolder}{output folder}

\item{colorRampPal}{colorRampPalette}

\item{zlims}{value limits for ploting}

\item{xlims}{x limits for ploting}

\item{ylims}{y limits for ploting}

\item{flip}{boolean, default=FALSE, flip the values of a Raster* object by inverting the order of the rows (direction=y)}
}
\value{
plot in pdf.
}
\description{
\code{help_plotNClayer} is plotting a single NetCDF file layer.
}
\examples{

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
