% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_summarySE.R
\name{help_summarySE}
\alias{help_summarySE}
\title{help_summarySE}
\usage{
help_summarySE(data = NULL, measurevar, groupvars = NULL,
  na.rm = FALSE, conf.interval = 0.95, .drop = TRUE)
}
\arguments{
\item{data}{data frame, Default: NULL}

\item{measurevar}{the name of a column that contains the variable to be summariezed}

\item{groupvars}{a vector containing names of columns that contain grouping variables, Default: NULL}

\item{na.rm}{a boolean that indicates whether to ignore NA's, Default: FALSE}

\item{conf.interval}{the percent range of the confidence interval, Default: 0.95}

\item{.drop}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
data frame
}
\description{
Gives count, mean, standard deviation, standard error of the mean, and confidence interval (default 95%).
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
