% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_ts2netcdf.R
\name{help_ts2netcdf}
\alias{help_ts2netcdf}
\title{Create netCDF from zoo time series object.}
\usage{
help_ts2netcdf(ts, unit, latitude, longitude, name.nc)
}
\arguments{
\item{ts}{zoo time series object, named.}
}
\value{
NULL, saves netCDF file containing all variables of zoo time series in working directory
}
\description{
Create netCDF from zoo time series object.
}
\examples{
library(zoo)

# dummy time series
date <- seq(as.Date("2010-01-01"),as.Date("2016-01-01"),1)
ts_data <- cbind(rnorm(length(date)), rnorm(length(date)))
colnames(ts_data) <- c("A","B")

ts_date  <- zoo(ts_data, date)
ts_POSIX <- zoo(ts_data, as.POSIXct(date))
ts_chron <- zoo(ts_data, as.chron(date))

# dummy lat, lon matrix
lat = as.matrix(40, ncol=1); dimnames(lat) = list("x","y")
lon = as.matrix(5 , ncol=1); dimnames(lat) = list("x","y")

help_ts2netcdf(ts = ts_date, unit = c("UNIT varA","UNIT varB"), 
               latitude = lat, longitude = lon, name.nc="test_date")
               
help_ts2netcdf(ts = ts_POSIX, unit = c("UNIT varA","UNIT varB"), 
               latitude = lat, longitude = lon, name.nc="test_POSIX")
               
help_ts2netcdf(ts = ts_chron, unit = c("UNIT varA","UNIT varB"), 
               latitude = lat, longitude = lon, name.nc="test_chron")

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
