% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_updateGITlib.R
\name{help_updateGITlib}
\alias{help_updateGITlib}
\title{Update GitHub library.}
\usage{
help_updateGITlib(lib, user = "JBrenn", branch = "master",
  bld_vignettes = FALSE)
}
\arguments{
\item{lib}{character, GitHub repository to update.}

\item{user}{character, GitHub user name.}

\item{branch}{character, GitHub repository branch}

\item{bld_vignettes}{boolean, build vignettes? default = FALSE}
}
\value{

}
\description{
Update GitHub library, by default from user repository "JBrenn"
}
\examples{

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
