% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_zooAggregation.R
\name{help_zooAggregation}
\alias{help_zooAggregation}
\title{Aggregation for zoo object with chron datetime representation.}
\usage{
help_zooAggregation(data, aggr_days = 1/24, na.rm)
}
\arguments{
\item{aggr_days}{time over which to aggregate in days, by default 1 hour (1/24 days).}

\item{x}{zoo object with chron datetime representation.}
}
\value{
aggregated zoo object
}
\description{
Aggregation for zoo object with chron datetime representation.
}
\examples{

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
