% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_v.R
\name{normalize_v}
\alias{normalize_v}
\title{min-max normalization}
\usage{
normalize_v(x)
}
\arguments{
\item{x}{data vector}
}
\value{
normalised data vector
}
\description{
min-max normalization of vector
}
\examples{
normalize_v(1:10)
}
