% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_rotatedAxisElementText.R
\name{rotatedAxisElementText}
\alias{rotatedAxisElementText}
\title{Return Element Text Object}
\usage{
rotatedAxisElementText(angle, position = "x")
}
\arguments{
\item{angle}{PARAM_DESCRIPTION}

\item{position}{PARAM_DESCRIPTION, Default: 'x'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Return Element Text Object in pecific angles
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
