% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_loadCarboEU.R
\name{help_loadCarboEU}
\alias{help_loadCarboEU}
\title{Load CarboEurope station data.}
\usage{
help_loadCarboEU(datapath, station)
}
\arguments{
\item{datapath}{full path to pre-processed data files}

\item{station}{station descrption, e.g. "ESES1"}
}
\value{
zoo object, containing all CarboEU station data, chron datetime representation
}
\description{
Load CarboEurope station data.
}
\examples{

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
