% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_plotNClayer.R
\name{help_plotNClayer}
\alias{help_plotNClayer}
\title{impute timeseries in netCDF (at site).}
\usage{
help_plotNClayer(infile, var = "pre", shp = NA, proj_utm = FALSE,
  crs_proj4 = "+proj=utm +zone=30 +ellps=intl +units=m +no_defs",
  plot_proj_utm = TRUE, outfolder,
  colorRampPal = colorRampPalette(brewer.pal(9, "Blues"))(100))
}
\arguments{
\item{infile}{netCDF file.}

\item{shp}{shape file, optional.}

\item{proj_utm}{boolean, is ncfile projected on UTM}

\item{crs_proj4}{character, proj4 projection definition}

\item{plot_proj_utm}{boolean , should out plot be projected}

\item{outfolder}{output folder}

\item{colorRampPal}{colorRampPalette}
}
\value{
plot in pdf.
}
\description{
\code{help_plotNClayer} is plotting a single NetCDF file layer.
}
\examples{

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
