% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_rotatedAxisElementText.R
\name{rotatedAxisElementText}
\alias{rotatedAxisElementText}
\title{Return Element Text Object}
\usage{
rotatedAxisElementText(angle, position = "x")
}
\arguments{
\item{angle}{angle in which to turn element [0;360]}

\item{position}{axis position, Default: 'x'}
}
\description{
Return Element Text Object in pecific angles
}
\examples{
\dontrun{
if(interactive()){
#Load Required Libraries
library(ggplot2)
library(gridExtra)
 #Demonstrate Usage for a Variety of Rotations
df    = data.frame(x=0.5,y=0.5)
plots = lapply(seq(0,90,length.out=4),function(a){
 ggplot(df,aes(x,y)) +
   geom_point() +
   theme(axis.text.x = rotatedAxisElementText(a,'x'),
         axis.text.y = rotatedAxisElementText(a,'y')) +
   labs(title = sprintf("Rotated \%s",a))
})
grid.arrange(grobs=plots)
 }
}
}
